# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API laravel 11",
        "version": "0.1"
    },
    "paths": {
        "/api/products": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "description": "Create a new product with the specified details.",
                "operationId": "index",
                "requestBody": {
                    "description": "Product details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "price",
                                    "stock",
                                    "timestamp"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Product name",
                                        "type": "string",
                                        "example": "Product Name"
                                    },
                                    "description": {
                                        "description": "Product description",
                                        "type": "string",
                                        "example": "Product Description"
                                    },
                                    "price": {
                                        "description": "Product price",
                                        "type": "number",
                                        "format": "float",
                                        "example": 10.99
                                    },
                                    "stock": {
                                        "description": "Product stock",
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "timestamp": {
                                        "description": "Timestamp",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-05-06 12:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "name": [
                                                    "The name field is required."
                                                ],
                                                "description": [
                                                    "The description field is required."
                                                ],
                                                "price": [
                                                    "The price field is required."
                                                ],
                                                "stock": [
                                                    "The stock field is required."
                                                ],
                                                "timestamp": [
                                                    "The timestamp field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product by ID",
                "description": "Retrieve the product information by its ID.",
                "operationId": "getProductById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "product not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update an existing product",
                "description": "Update an existing product with the specified details.",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Product details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Product name",
                                        "type": "string",
                                        "example": "Updated Product Name"
                                    },
                                    "description": {
                                        "description": "Product description",
                                        "type": "string",
                                        "example": "Updated Product Description"
                                    },
                                    "price": {
                                        "description": "Product price",
                                        "type": "number",
                                        "format": "float",
                                        "example": 19.99
                                    },
                                    "stock": {
                                        "description": "Product stock",
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "timestamp": {
                                        "description": "Timestamp",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-05-06 12:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "name": [
                                                    "The name field is required."
                                                ],
                                                "description": [
                                                    "The description field is required."
                                                ],
                                                "price": [
                                                    "The price field is required."
                                                ],
                                                "stock": [
                                                    "The stock field is required."
                                                ],
                                                "timestamp": [
                                                    "The timestamp field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a product",
                "description": "Delete a product based on the provided ID.",
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Product deleted successfully"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "title": "Product",
                "description": "Product model",
                "required": [
                    "name",
                    "description",
                    "price",
                    "stock"
                ],
                "properties": {
                    "name": {
                        "description": "The name of the product",
                        "type": "string"
                    },
                    "description": {
                        "description": "The description of the product",
                        "type": "string"
                    },
                    "price": {
                        "description": "The price of the product",
                        "type": "number",
                        "format": "float"
                    },
                    "stock": {
                        "description": "The stock quantity of the product",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    }
}